test(api): migrate billing redeem code to api-first bdd
Round 30 of the API-first BDD migration (#16967), BILLING REDEEM CODE.

- Extend createBddApi with the billingRedeemCode client.
- Add zero-billing-redeem-code.bdd.test.ts: an admin redeems through the external
  Atom service (MSW-mocked) with a Clerk M2M token, asserting the outbound Atom
  request (trimmed code + org id), the M2M params, and {redeemed:true}; non-admin
  / unauthenticated callers never reach Atom; missing config, M2M failure, and an
  unreachable Atom surface as 503; every Atom rejection shape maps to its stable
  bad-request message.
- Delete zero-billing-redeem-code.test.ts whole. Also document
  webhooks-built-in-generations.test.ts as a pure-function service exception.

Coverage verified (source-only): zero-billing-redeem-code.ts (70/42/9) and
zero-billing-redeem.service.ts (89/31/7) unchanged vs main. No regressions.
