+++ turbo/apps/api/src/signals/routes/__tests__/helpers/api-bdd.ts
@@ -19,6 +19,7 @@ import {
 } from "@vm0/api-contracts/contracts/zero-personal-model-providers";
 import { zeroUserConnectorsContract } from "@vm0/api-contracts/contracts/user-connectors";
 import { zeroOrgListContract } from "@vm0/api-contracts/contracts/zero-org-list";
+import { zeroBillingRedeemCodeContract } from "@vm0/api-contracts/contracts/zero-billing";
 import {
   zeroOrgInviteContract,
   zeroOrgMembershipRequestsContract,
@@ -173,6 +174,8 @@ export function createBddApi(context: TestContext) {
     membershipRequests: setupApp({ context })(
       zeroOrgMembershipRequestsContract,
     ),
+    /** ts-rest client for `/api/zero/billing/redeem-code` (consume a code). */
+    billingRedeemCode: setupApp({ context })(zeroBillingRedeemCodeContract),
     /** ts-rest client for `/api/zero/attribution/signup` (record first-touch
      * signup attribution into Clerk private metadata). */
     attribution: setupApp({ context })(zeroAttributionContract),
